Minor QoL fixes and introducing better logic for disenchanting automatically

+++ src/main/kotlin/league/models/json/ChallengeInfo.kt
@@ -47,7 +47,7 @@ class ChallengeInfo {
 
     val thresholdSummary get() = try {
         thresholds!!.toList().sortedBy { it.first }.filter { it.first > nextLevel!! }
-            .joinToString(" > ") { it.second.value!!.toInt().toReadableNumber() }
+            .joinToString(" > ") { it.second.value!!.toLong().toReadableNumber() }
     } catch (_: Exception) {
         ""
     }
